refactor: 3層アーキテクチャにリファクタリング
- セッション作成のビジネスロジックをサービス層に分離
- API層はHTTPハンドリングのみに責務を限定
- アーキテクチャ設計をドキュメント化
